Feature ETP-81: Add cache to menu metadata features.
